reorganizing database interop code
