+++ Project.toml
@@ -8,6 +8,8 @@ version = "0.2.21"
 AlgebraicInterfaces = "23cfdc9f-0504-424a-be1f-4892b28e2f0c"
 Base64 = "2a0f44e3-6c83-55bd-87e4-b1978d98bd5f"
 CompTime = "0fb5dd42-039a-4ca4-a1d7-89a96eae6d39"
+DBInterface = "a10d1c49-ce27-4219-8d33-6db1a4562965"
+DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
 MLStyle = "d8e11817-5142-5d16-987a-aa16d5891078"
@@ -25,17 +27,27 @@ StructTypes = "856f2bd8-1eba-4b0a-8007-ebc267875bd4"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [weakdeps]
+DuckDB = "d2f5444f-75bc-4fdf-ac35-56f514c445e1"
+LibPQ = "2196b101-3ac0-4e12-b5a8-532424305d17"
+MySQL = "39abe10b-433b-5dbd-92d4-e302a9df00cd"
+SQLite = "0aa819cd-b072-5ff4-a722-6bc24af294d9"
 XLSX = "fdbf4ff8-1666-58a4-91e7-1b58723a45e0"
 nauty_jll = "55c6dc9b-343a-50ca-8ff2-b71adb3733d5"
 
 [extensions]
+DuckDBACSetsExt = "DuckDB"
+LibPQACSetsExt = "LibPQ"
+MySQLACSetsExt = "MySQL"
 NautyACSetsExt = "nauty_jll"
+SQLiteACSetsExt = "SQLite"
 XLSXACSetsExt = "XLSX"
 
 [compat]
 AlgebraicInterfaces = "0.1.4"
 Base64 = "1.9"
 CompTime = "0.1"
+DBInterface = "2.6.1"
+DataFrames = "1.7.0"
 DataStructures = "0.18"
 JSON3 = "1.13.2"
 MLStyle = "0.4"
